Dev version v0.5
Game is, logic wise, finished. It has a start state, fail state and a won state. As well as a way to reset when either fail and won has reached. This version will be pushed to main as beta v0.1

Next up is adding visuals to indicate damage state. This is visualised by a boat breaking. Also on the planning is adding sound.