Site update
To showcase the newer version of paddle, I updated the website. Now it is more inline with the play rules document. Concept video is still part of it, next to a newer explenation what the game is about. I also added a link to the playable prototype and a link to the play rules PDF. It needs some refinement before it is merged with main.